services/agentic-recommendation-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
motor==3.3.2
pydantic==2.5.3
pymongo==4.6.1
python-dotenv==1.0.0
kafka-python==2.0.2
pandas==2.1.4
python-multipart==0.0.6

